+++ ReleaseNotes.md
@@ -5,6 +5,8 @@
 * add start and runntime info to tasklist
 * add failed state (exclamation mark)
 * add colors to state icons
+* add setup steps for custom profiles
+* redesign process pipes to be more robust
 
 ## 0.00.001