+++ ReleaseNotes.md
@@ -3,6 +3,8 @@
 ## 0.00.003
 
 **features:**
+* [x] add floating button to scroll down and copy stdout
+* [x] improve auto scroll behavior
 
 **bug-fixes:**
 * [x] prevent app freeze by limiting console string length
@@ -10,7 +12,7 @@
 ## 0.00.002
 
 **features:**
-* [x] add start and runtime info to tasklist
+* [x] add start and runtime info to task-list
 * [x] add failed state (exclamation mark)
 * [x] add colors to state icons
 * [x] add setup steps for custom profiles